cleaned up Auth Setup + added Postman export

+++ sql/database/querier.go
@@ -10,7 +10,7 @@ import (
 
 type Querier interface {
 	CreateUser(ctx context.Context, arg CreateUserParams) (*User, error)
-	GetUserById(ctx context.Context, id int64) (*User, error)
+	GetUserById(ctx context.Context, id int32) (*User, error)
 	GetUserByName(ctx context.Context, username string) (*User, error)
 }
